VAAPI transcoding code refactoring
VAAPI transcoding code refactoring to match the ffmpeg/docs/example recomandations.
