+++ Makefile
@@ -531,6 +531,9 @@ endif
 ifeq ($(CONFIG_LIBOPUS),yes)
 LIBS-CODECS += libopus
 endif
+ifeq ($(CONFIG_VAAPI_OLD),yes)
+LIBS-CODECS += vaapi
+endif
 ifeq ($(CONFIG_VAAPI),yes)
 LIBS-CODECS += vaapi
 endif
@@ -545,6 +548,9 @@ SRCS-CODECS += $(foreach lib,$(LIBS-CODECS),src/transcoding/codec/codecs/libs/$(
 #hwaccels
 ifeq ($(CONFIG_HWACCELS),yes)
 SRCS-HWACCELS += src/transcoding/transcode/hwaccels/hwaccels.c
+ifeq ($(CONFIG_VAAPI_OLD),yes)
+SRCS-HWACCELS += src/transcoding/transcode/hwaccels/vaapi_old.c
+endif
 ifeq ($(CONFIG_VAAPI),yes)
 SRCS-HWACCELS += src/transcoding/transcode/hwaccels/vaapi.c
 endif
