+++ src/transcoding/codec.h
@@ -171,7 +171,7 @@ htsmsg_t *
 codec_get_profiles_list(enum AVMediaType media_type);
 
 void
-#if ENABLE_VAAPI
+#if ENABLE_VAAPI_OLD || ENABLE_VAAPI
 codec_init(int vainfo_probe_enabled);
 #else
 codec_init(void);
