+++ src/transcoding/transcode/hwaccels/hwaccels.h
@@ -51,12 +51,21 @@ hwaccels_get_sharpness_filter(AVCodecContext *avctx, int value, char *filter, si
 
 
 /* encoding ================================================================= */
+int
+hwaccels_initialize_encoder_from_decoder(AVCodecContext *iavctx, AVCodecContext *oavctx);
 
 int
+#if ENABLE_VAAPI
+hwaccels_encode_setup_context(AVCodecContext *avctx);
+#else
 hwaccels_encode_setup_context(AVCodecContext *avctx, int low_power);
+#endif
 
 void
 hwaccels_encode_close_context(AVCodecContext *avctx);
+
+enum AVPixelFormat
+hwaccels_get_pixfmt_format_for_filter(AVCodecContext *avctx);
 
 
 /* module =================================================================== */
